+++ components/ADempiere/DefaultTable/Components/ColumnsDisplayOption.vue
