+++ components/ADempiere/DefaultTable/Components/IndexColumn.vue
