+++ components/ADempiere/FieldDefinition/FieldSearch/GeneralInfoSearch/panel.vue
@@ -137,10 +137,10 @@ import { DISPLAY_COLUMN_PREFIX } from '@/utils/ADempiere/dictionaryUtils'
 
 // components and mixins
 import fieldSearchMixin from '../mixinFieldSearch'
-import CellInfo from '@theme/components/ADempiere/DefaultTable/CellInfo.vue'
+import CellInfo from '@theme/components/ADempiere/DefaultTable/Components/CellInfo.vue'
 import FieldDefinition from '@theme/components/ADempiere/FieldDefinition/index.vue'
-import CustomPagination from '@theme/components/ADempiere/DefaultTable/CustomPagination.vue'
-import IndexColumn from '@theme/components/ADempiere/DefaultTable/IndexColumn.vue'
+import CustomPagination from '@theme/components/ADempiere/DefaultTable/Components/CustomPagination.vue'
+import IndexColumn from '@theme/components/ADempiere/DefaultTable/Components/IndexColumn.vue'
 
 // utils and helper methods
 import { isEmptyValue, isSameValues } from '@/utils/ADempiere/valueUtils'
